backend/requirements.txt
# Mental Health Chatbot — Python Dependencies
# Install with: pip install -r requirements.txt
#
# NOTE: If you are on Python 3.14 and psycopg2-binary fails,
# run this instead: pip install psycopg2-binary --pre

# PDF processing
pymupdf>=1.24.0

# Token-based chunking (RecursiveCharacterTextSplitter)
langchain-text-splitters>=0.3.0
tiktoken>=0.7.0

# Vector database
qdrant-client>=1.9.0

# PostgreSQL driver
psycopg2-binary>=2.9.9

# Supabase (PostgreSQL backend-as-a-service)
supabase>=2.11.0

# Ollama Python client
ollama>=0.2.1

# API server
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.7.1

# Task scheduling — daily synthetic check-in/wearable refresh
apscheduler>=3.10.4

# Utilities
tqdm>=4.66.4
python-dotenv>=1.0.1
